@@ -1,8 +1,11 @@
 <?php
 namespace sbronsted;
 
+/**
+ * Class ConsoleWriter writes it's output on stdout
+ */
 class ConsoleWriter implements LogWriter {
-	public function write($level, $class, $text) {
+	public function write(string $level, string $class, string $text) : void {
 		print("$level:$class: $text\n");
 	}
 }
